Removes files and codes not used anymore
